+++ lib/resource_policy.rb
@@ -6,6 +6,7 @@ require 'resource_policy/version'
 module ResourcePolicy
   class Error < StandardError; end
 
+  require 'resource_policy/configuration'
   require 'resource_policy/policy'
   # Your code goes here...
 end